requirements.txt
alembic==1.8.1
bcrypt==3.2.2
cryptography==37.0.4
email-validator==1.2.1
fastapi==0.65.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
loguru==0.6.0
passlib==1.7.4
passlib[bcrypt]
psutil==5.9.1
pydantic==1.5.1
PyJWT==1.7.1
PyMySQL==1.0.2
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
SQLAlchemy==1.3.17
uvicorn==0.18.2
zipp==3.8.1